backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "bs4>=0.0.2",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "greenlet>=3.2.3",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pypdf>=5.8.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "tavily-python>=0.7.9",
    "uvicorn>=0.35.0",
]
